+++ Define.h
@@ -2,3 +2,4 @@
 
 #define SQOMDLI_VER "0.5"
 #define SAOMDLI_RELEASE __DATE__
+#define DEMO_VIDEO_URL "https://youtu.be/ockk6MTUR38"